Move SignUp and SignIn components to a subdirectory
