+++ web/libpointmatcher-server-interface/src/components/Auth/SignIn.vue
@@ -45,8 +45,8 @@
   
 
   <script>
-  import EyeOpenIcon from './icons/IconEyeOpen.vue'
-  import EyeSlashIcon from './icons/IconEyeSlash.vue'
+  import EyeOpenIcon from '../icons/IconEyeOpen.vue'
+  import EyeSlashIcon from '../icons/IconEyeSlash.vue'
   export default {
     components: {
       EyeOpenIcon,
